Uptade Dado and Model class
